requirements.txt
beautifulsoup4==4.6.0
requests==2.18.4
py-postgresql==1.2.1
flask==0.12.2